% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{ascends}
\alias{ascends}
\title{Determine if cell locations are stored in ascending or descending order.}
\usage{
ascends(x, dir = "lat")
}
\arguments{
\item{x}{netcd4 or raster object}

\item{dir}{character either 'lon' or 'lat' (default)  d}
}
\value{
logical, TRUE if the specified cell locations ascend in value
}
\description{
Determine if cell locations are stored in ascending or descending order.
}
