% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{bbox_to_polygon}
\alias{bbox_to_polygon}
\title{Convert a bbox to sp::polygon, sp::Polygons, or sp::SpatialPolygons object}
\usage{
bbox_to_polygon(bb, projstring = "+proj=longlat +datum=WGS84",
  id = "bbox", output_class = c("SpatialPolygons", "Polygons",
  "Polygon")[1])
}
\arguments{
\item{bb}{numeric, a 4-element bbox vector [left, right, bottom, top]}

\item{projstring}{character sutiable to pass to \code{sp::CRS}, by default "+proj=longlat +datum=WGS84"}

\item{id}{character, the polygon ID, by default 'bbox'}

\item{output_class}{character, either "SpatialPolygons", "Polygons" or "Polygon"}
}
\value{
sp::SpatialPolygons object or NULL
}
\description{
Convert a bbox to sp::polygon, sp::Polygons, or sp::SpatialPolygons object
}
