% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlendedSeaWindRefClass.R
\docType{class}
\name{BlendedSeaWindsRefClass-class}
\alias{BlendedSeaWindsRefClass}
\alias{BlendedSeaWindsRefClass-class}
\title{A subclass of SPNCRefClass for Blended Sea Winds}
\description{
A subclass of SPNCRefClass for Blended Sea Winds
}

