% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{coerce_within}
\alias{coerce_within}
\title{Coerce values of x two be within the range of the bounds b.}
\usage{
coerce_within(x, b = c(0, 1))
}
\arguments{
\item{x}{numeric vector}

\item{b}{two element numeric vector of [min,max] defining bounds}
}
\value{
a numeric vector same length as x with all values within bounds b
}
\description{
Coerce values of x two be within the range of the bounds b.
}
