% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPCUGBDP.R
\name{CPCUGBDP_get_points}
\alias{CPCUGBDP_get_points}
\title{Get points for CPCUGBDPRefClass}
\usage{
CPCUGBDP_get_points(X, x, y = NULL, what = X$VARS[1], layer = 1)
}
\arguments{
\item{X}{CPCUGBDPRefClass object}

\item{x}{numeric, vector of longitude points or, if y is NULL, a matrix [x,y] 
or a data.frame or list with x,y elements}

\item{y}{numeric vector of lattitude points or NULL}
}
\value{
numeric vector of values
}
\description{
Get points for CPCUGBDPRefClass
}

