% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CPCUGBDP.R
\name{CPCUGBDPRefClass_step}
\alias{CPCUGBDPRefClass_step}
\title{Get the step size (resolution) as a guess - the space between the first
on and lat values.}
\value{
two element numeric vector of step size in x and y or NULL
These may be signed for descending values of lon and or lat
}
\description{
Get the step size (resolution) as a guess - the space between the first
on and lat values.
}

