% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CPCUGBDP.R
\name{CPCUGBDPRefClass_subset_bbox}
\alias{CPCUGBDPRefClass_subset_bbox}
\title{Craft subset indices into a ncdf array}
\arguments{
\item{bb}{numeric, four element bounding box [left, right, bottom, top]}

\item{time}{numeric time to retrieve}
}
\value{
a list of \code{start} indices in x and y, \code{counts} in x and y and
   a possibly updated copy of \code{bb} vector of [left, right, bottom, top]
}
\description{
Craft subset indices into a ncdf array
}
\seealso{
\url{http://r.789695.n4.nabble.com/Lat-Lon-NetCDF-subset-td3394326.html}
}

