% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_matrix}
\alias{get_matrix}
\title{Extract a matrix of data from a SPNCRefClass of raster flavor}
\usage{
get_matrix(x, var, layer = 1, ...)
}
\arguments{
\item{x}{the SPNCRefClass or subclass object}

\item{var}{character, the name of the variable}

\item{layer}{index of the layer (by default the first)}

\item{...}{further arguments for SPNC$get_raster() method}
}
\value{
a numeric matrix or NULL
}
\description{
Extract a matrix of data from a SPNCRefClass of raster flavor
}
