% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_projection}
\alias{get_projection}
\title{Retrieve projection CRS by name}
\usage{
get_projection(name = c("longlat", "eqc"))
}
\arguments{
\item{name}{character, the name of the projection}
}
\value{
character projection string (proj4)
}
\description{
Retrieve projection CRS by name
}
