% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{get_shape}
\alias{get_shape}
\title{Get a 2 element vector of [nx, ny] dimensions}
\usage{
get_shape(x)
}
\arguments{
\item{x}{netcd4 or raster object}
}
\value{
a two element vector of [nx, ny]
}
\description{
Get a 2 element vector of [nx, ny] dimensions
}
