% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HMODISL3SMI.R
\docType{class}
\name{HMODISL3SMIRefClass-class}
\alias{HMODISL3SMIRefClass}
\alias{HMODISL3SMIRefClass-class}
\title{A subclass of SPNCRefClass for Modis Aqua via \url{http://thredds.jpl.nasa.gov}}
\description{
Warning - metadata looks corrupted to me
}

