% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HMODISL3SMI.R
\name{HMODISL3SMIRefClass_get_global_atts}
\alias{HMODISL3SMIRefClass_get_global_atts}
\title{Get global attributes}
\value{
a named list of global attributes or NULL
}
\description{
Get global attributes
}

