% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HMODISL3SMI.R
\name{HMODISL3SMIRefClass_subset_coords}
\alias{HMODISL3SMIRefClass_subset_coords}
\title{Retrieve the subset coordinates}
\arguments{
\item{bb}{4 element vector in standard form [left, right, bottom, top] [-180,180]}

\item{lon}{numeric vector of lons (ascending order please!) to select from}

\item{lat}{numeric vector of lats (ditto)}
}
\value{
a list of \code{start} indices in x and y, \code{counts} in x and y and
   a possibly updated copy of \code{bb} vector of [left, right, bottom, top]
}
\description{
Retrieve the subset coordinates
}

