% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlendedSeaWindRefClass.R
\name{is_BSW}
\alias{is_BSW}
\title{Test if an NCDF contains Blended Sea Winds data.}
\usage{
is_BSW(x)
}
\arguments{
\item{x}{ncdf4 object or SPNCRefClass}
}
\value{
logical
}
\description{
Test if an NCDF contains Blended Sea Winds data.
}

