% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPCUGBDP.R
\name{is_CPCUGBDP}
\alias{is_CPCUGBDP}
\title{Test if an NCDF contains CPCUGBDP data.}
\usage{
is_CPCUGBDP(x)
}
\arguments{
\item{x}{ncdf4 object or SPNCRefClass}
}
\value{
logical
}
\description{
Test if an NCDF contains CPCUGBDP data.
}
