% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3SMI.R
\name{is_L3SMI}
\alias{is_L3SMI}
\title{Test if an NCDF contains L3SMI data.}
\usage{
is_L3SMI(x)
}
\arguments{
\item{x}{ncdf4 object or SPNCRefClass}
}
\value{
logical
}
\description{
Test if an NCDF contains L3SMI data.
}
