% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MURSST.R
\name{is_MURSST}
\alias{is_MURSST}
\title{Test if an NCDF contains MURSST data.}
\usage{
is_MURSST(x)
}
\arguments{
\item{x}{ncdf4 object or SPNCRefClass}
}
\value{
logical
}
\description{
Test if an NCDF contains MURSST data.
}

