% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAMANL.R
\name{is_NAMANL}
\alias{is_NAMANL}
\title{Test if an NCDF contains NAM-ANL data.}
\usage{
is_NAMANL(x)
}
\arguments{
\item{x}{ncdf4 object or NAMANLRefClass}
}
\value{
logical
}
\description{
Test if an NCDF contains NAM-ANL data.
}

