% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NARR.R
\name{is_NARR}
\alias{is_NARR}
\title{Test if an NCDF contains NARR data.}
\usage{
is_NARR(x)
}
\arguments{
\item{x}{ncdf4 object or SPNCRefClass}
}
\value{
logical
}
\description{
Test if an NCDF contains NARR data.
}

