% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCEP.R
\name{is_NCEP}
\alias{is_NCEP}
\title{Test if an NCDF contains NCEP/NMC data.}
\usage{
is_NCEP(x)
}
\arguments{
\item{x}{ncdf4 object or SPNCRefClass}
}
\value{
logical
}
\description{
Test if an NCDF contains NCEP/NMC data.
}
