% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OISST.R
\name{is_OISST}
\alias{is_OISST}
\title{Test if an NCDF contains OISST data.}
\usage{
is_OISST(x)
}
\arguments{
\item{x}{ncdf4 object or SPNCRefClass}
}
\value{
logical
}
\description{
Test if an NCDF contains OISST data.
}

