% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3SMI.R
\docType{class}
\name{L3SMIRefClass-class}
\alias{L3SMIRefClass}
\alias{L3SMIRefClass-class}
\title{A subclass of SPNCRefClass for OBPG L3 Standard Mapped Image}
\description{
A subclass of SPNCRefClass for OBPG L3 Standard Mapped Image
}

