% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3SMI.R
\name{L3SMIRefClass_get_extent}
\alias{L3SMIRefClass_get_extent}
\title{Compute extent given a bounding box}
\arguments{
\item{bb}{the bounding box needed, if missing the current one is used}
}
\value{
a raster::Extent object
}
\description{
Compute extent given a bounding box
}
