% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3SMI.R
\name{L3SMIRefClass_step}
\alias{L3SMIRefClass_step}
\title{Get the step size (resolution) as reported in the global attributes.}
\value{
two element numeric vector of step size in x and y or NULL.
}
\description{
Get the step size (resolution) as reported in the global attributes.
}

