% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mgrepl}
\alias{mgrepl}
\title{Perform grepl on multiple patterns; it's like OR-ing successive grepl statements.}
\usage{
mgrepl(pattern, x, ...)
}
\arguments{
\item{pattern}{character vector of patterns}

\item{x}{the character vector to search}

\item{...}{further arguments for \code{grepl}}
}
\value{
logical vector of matches where at least one element of \code{pattern}
   is found in \code{x}
}
\description{
Perform grepl on multiple patterns; it's like OR-ing successive grepl statements.
}
