% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MODISL3SMI.R
\docType{class}
\name{MODISL3SMIRefClass-class}
\alias{MODISL3SMIRefClass}
\alias{MODISL3SMIRefClass-class}
\title{A subclass of SPNCRefClass for Modis Aqua \url{http://oceancolor.gsfc.nasa.gov/cms/}}
\description{
A subclass of SPNCRefClass for Modis Aqua \url{http://oceancolor.gsfc.nasa.gov/cms/}
}

