% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MURSST.R
\name{MURSST_get_points}
\alias{MURSST_get_points}
\title{Get points for MURSSTRefClass}
\usage{
MURSST_get_points(NC, x, y = NULL, what = NC$VARS[1], layer = 1)
}
\arguments{
\item{NC}{MURSSTRefClass object}

\item{x}{numeric, vector of longitude points or, if y is NULL, a matrix [x,y]
or a data.frame or list with x,y elements}

\item{y}{numeric vector of lattitude points or NULL}
}
\value{
numeric vector of values
}
\description{
Get points for MURSSTRefClass
}

