% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MURSST.R
\docType{class}
\name{MURSSTRefClass-class}
\alias{MURSSTRefClass}
\alias{MURSSTRefClass-class}
\title{A subclass of SPNCRefClass for Multi-scale Ultra-high Resolution Sea Surface Temperature \url{http://mur.jpl.nasa.gov/}}
\description{
A subclass of SPNCRefClass for Multi-scale Ultra-high Resolution Sea Surface Temperature \url{http://mur.jpl.nasa.gov/}
}

