% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MURSST.R
\name{MURSSTRefClass_get_points}
\alias{MURSSTRefClass_get_points}
\title{Get points}
\arguments{
\item{x}{numeric, vector of longitude points or, if y is NULL, a matrix [x,y]
or a data.frame or list with x,y elements}

\item{y}{numeric vector of lattitude points or NULL}

\item{what}{character one or more variable names or variable indices}
}
\value{
numeric vector of values
}
\description{
Get points
}

