% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MURSST.R
\name{MURSSTRefClass_subset_bbox}
\alias{MURSSTRefClass_subset_bbox}
\title{Craft subset indices into a ncdf array}
\arguments{
\item{bb}{numeric, four element bounding box [left, right, bottom, top]}
}
\value{
a list of
\itemize{
 \item{start indices in x and y}
 \item{counts in x and y}
 \item{bb vector of [left, right, bottom, top] pixel centers}
 \item{ext extent vector [left, right, bottom, top] pixel outer edges}
 }
}
\description{
Craft subset indices into a ncdf array
}
\seealso{
\url{https://stat.ethz.ch/pipermail/r-help/2011-March/272641.html}
}
