% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MURSST.R
\name{MURSSTRefClass_subset_points}
\alias{MURSSTRefClass_subset_points}
\title{Compute indicies (start and count) for individual [lon,lat] points}
\arguments{
\item{x}{either a vector of longitude or a 2d matrix [lon,lat] or a data.frame
with lon,lat columns}

\item{y}{numeric vector, if x is a vector, otherwise NULL}
}
\value{
a list of start/count elements, one per xy input.  For inputs that
   fall beyond the bounds of the data then NA is returned.
}
\description{
Compute indicies (start and count) for individual [lon,lat] points
}
