% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAMANL.R
\name{NAMANL_get_raster}
\alias{NAMANL_get_raster}
\title{Get a raster for NAMANLRefClass}
\usage{
NAMANL_get_raster(NC, what = NC$VARS[1], bb = NC$BB,
  crs = "+proj=longlat +datum=WGS84", flip = FALSE, time_fmt = "D\%Y\%j",
  ...)
}
\arguments{
\item{NC}{NAMANLRefClass object}

\item{what}{character one or more variable names or variable indices}

\item{bb}{a 4 element bounding box vector [left, right, bottom, top], defaults
to [-180, 180, -90, 90]}

\item{crs}{character, the coordinate reference system to apply}

\item{flip}{logical if TRUE then flip the raster in the y direction}

\item{time_fmt}{if multiple time layers are returned, this controls the layer names}

\item{...}{further arguments to select the parameter by name}
}
\value{
a \code{raster::brick} or \code{raster::layer} object or NULL
}
\description{
Get a raster for NAMANLRefClass
}
