% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAMANL.R
\docType{class}
\name{NAMANLRefClass-class}
\alias{NAMANLRefClass}
\alias{NAMANLRefClass-class}
\title{A subclass of SPNCRefClass for NAM-ANL}
\description{
A subclass of SPNCRefClass for NAM-ANL
}
\section{Fields}{

\describe{
\item{\code{lambert}}{a Raster* class object}

\item{\code{mercator}}{a Raster* class object}
}}

