% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAMANL.R
\name{NAMANLRefClass_get_extent}
\alias{NAMANLRefClass_get_extent}
\title{Compute extent given a bounding box}
\arguments{
\item{bb}{the bounding box needed, if missing the current one is used, always 
assumed to be longlat}

\item{proj}{character string either "lcc" or "longlat" (default) to specify the 
projection of the returned}
}
\value{
a raster::Extent object
}
\description{
Compute extent given a bounding box
}
