% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NARR.R
\name{NARR_get_points}
\alias{NARR_get_points}
\title{Get data values by specifying long/lat locations}
\usage{
NARR_get_points(X, x, y = NULL, time = 1, ...)
}
\arguments{
\item{X}{NARRRefClass object}

\item{x}{numeric vector of x locations (longitude), or a matrix, data frame or list
of x and y locations}

\item{y}{numeric vector of y locations - if NULL then these must be part of x}

\item{time}{numeric index or POSIXct}

\item{...}{further arguments for \code{raster::extract()}}
}
\value{
as specified by \code{raster::extract()}
}
\description{
Get data values by specifying long/lat locations
}
