% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NARR.R
\name{NARRREfClass_bbox_project}
\alias{NARRREfClass_bbox_project}
\title{Convert bb to longlat}
\arguments{
\item{what}{ignored}

\item{time}{numeric vector either a 1-based indices or POSIXct timestamps}

\item{bb}{a 4 element bounding box vector [left, right, bottom, top]
defaults to the extent}

\item{crs}{character, the coordinate reference system to apply}

\item{flip}{logical if TRUE then flip the raster in the y direction}
}
\value{
Retrieve a raster, possibly a cropped

a \code{raster::brick} or \code{raster::layer} object or NULL
}
\description{
Convert bb to longlat
}
