% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NARR.R
\name{NARRRefClass_get_extent}
\alias{NARRRefClass_get_extent}
\title{Compute extent given a bounding box}
\arguments{
\item{bb}{the bounding box needed, if missing the current one is used}

\item{as_longlat}{logical if TRUE inverse project the exent to long lat values}
}
\value{
a raster::Extent object
}
\description{
Compute extent given a bounding box
}
