% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NARR.R
\name{NARRRefClass_get_raster}
\alias{NARRRefClass_get_raster}
\alias{NARR_get_raster}
\title{Retrieve a raster, possibly a cropped}
\usage{
NARR_get_raster(X, bb = NULL, time = 1, what = NULL,
  crs = raster::projection(X$R), flip = FALSE)
}
\arguments{
\item{X}{NARRRefClass object}

\item{bb}{a 4 element bounding box vector [left, right, bottom, top], defaults
to [-180, 180, -90, 90]}

\item{time}{numeric vector either a 1-based indices or POSIXct timestamps}

\item{what}{ignored}

\item{crs}{character, the coordinate reference system to apply}

\item{flip}{logical if TRUE then flip the raster in the y direction}
}
\value{
a \code{raster::RasterBrick} or \code{raster::RasterLayer} object or NULL
}
\description{
Retrieve a raster, possibly a cropped
}

