% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NARR.R
\name{NARRRefClass_lat}
\alias{NARRRefClass_lat}
\title{Get the latitude values of the grid}
\arguments{
\item{what}{the desired location - 'center' or 'native' (default) which are the same.}
}
\value{
numeric vector or NULL
}
\description{
Get the latitude values of the grid
}

