% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NARR.R
\name{NARRRefClass_lon}
\alias{NARRRefClass_lon}
\title{Get the longitude values of the grid.  Note the grid is projected, so}
\arguments{
\item{what}{the desired location - 'center' or 'native' (default) which are the same.}

\item{what}{the desired location - 'center' or 'native' (default) which are the same.}
}
\value{
numeric vector or NULL

numeric vector or NULL
}
\description{
Get the longitude values of the grid.  Note the grid is projected, so

Get the longitude values of the grid
}
