% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NARR.R
\name{NARRRefClass_step}
\alias{NARRRefClass_step}
\title{Get the nominal step size (resolution) in km}
\value{
two element numeric vector of step size in x and y or NULL
}
\description{
Get the nominal step size (resolution) in km
}
\seealso{
http://www.nco.ncep.noaa.gov/pmb/docs/on388/tableb.html#GRID221
}

