% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{nc_crop}
\alias{nc_crop}
\title{Crop a ncdf4 object}
\usage{
nc_crop(x, varname = "sst", bb = c(-180, 180, -90, 90),
  template = nc_template(x))
}
\arguments{
\item{x}{ncdf4 object}

\item{varname}{character variable name 'sst' by default}

\item{bb}{4 element bounding box [left, right, bottom, top]}

\item{template}{raster template}
}
\value{
raster
}
\description{
Crop a ncdf4 object
}
