% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{nc_step}
\alias{nc_step}
\title{Compute the mean step.  If the step size is provided in the global attributes
then it is advised you use that information.  See \code{ncglobal_atts()}}
\usage{
nc_step(NC, dim_names = c("lon", "lat"))
}
\arguments{
\item{NC}{a ncdf4 class object}

\item{dim_names}{a two element character vector providing the names of the
x and y coordinates.  By default \code{c('lon', 'lat')}}
}
\value{
a 2 element numeric of the mean step size - always positive
}
\description{
Compute the mean step.  If the step size is provided in the global attributes
then it is advised you use that information.  See \code{ncglobal_atts()}
}
