% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{nc_subset}
\alias{nc_subset}
\title{Compute subset coordinates given a bounding box - suitable for use with
\code{ncvar_get()}.  This is sort of a generic function that may not work for
all NetCDF files. Coordinates within the files are assumed to be pixel centers.}
\usage{
nc_subset(NC, bb = c(-77, -51.5, 37.9, 56.7), step = nc_step(NC),
  dim_names = c("lon", "lat"))
}
\arguments{
\item{NC}{a ncdf4 class object}

\item{bb}{the 4 element bounding box [left, right, bottom, top]}

\item{step}{one or two element numeric of the step size.}

\item{dim_names}{a two element character vector providing the names of the
x and y coordinates.  By default \code{c('lon', 'lat')}}
}
\value{
a 5 element list of
 \itemize{
     \item{start a 2 element vector of indices to start}
     \item{count a 2 element vector number of columns/rows}
     \item{ext a 4 element extent that accounts for the pixels sizes - the envelope}
     \item{y_ascends logical TRUE if y values are stored low to high}
     \item{step two element step size (aka resolution)}
  }
}
\description{
Pixel center coordinates are typically specified within the 'dim' attribute in the
ncdf4 object.  Step size between pixel centers is taken to be the mean for
all pixels by default, but this can be overridden.
}
