% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{nc_template}
\alias{nc_template}
\title{Create a raster template from the contents of a ncdf4 object}
\usage{
nc_template(x, proj = get_projection("longlat"))
}
\arguments{
\item{x}{ncdf4 object}

\item{proj}{CRS projection string}
}
\value{
raster
}
\description{
Create a raster template from the contents of a ncdf4 object
}
