% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{ncdim_get}
\alias{ncdim_get}
\title{Retrieve a vector of dimensions}
\usage{
ncdim_get(NC)
}
\arguments{
\item{NC}{a ncdf4 object}
}
\value{
a named vector of dimesnions
}
\description{
Retrieve a vector of dimensions
}
