% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCEP.R
\docType{class}
\name{NCEPRefClass-class}
\alias{NCEPRefClass}
\alias{NCEPRefClass-class}
\title{A subclass of SPNCRefClass for NCEP/NCM}
\description{
This NetCDF data is projected and can be ready by the \code{raster} package. Unlike
other members of the SPNCRefClass, this class stores the data as raster.  The get
\code{get_raster()} and \code{get_points()} functions are still exposed for consistency.
}
\section{Fields}{

\describe{
\item{\code{R}}{a Raster* class object}
}}

