% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nc.R
\name{ncglobal_atts}
\alias{ncglobal_atts}
\title{Retrieve a named list of global attributes}
\usage{
ncglobal_atts(NC, rm_pattern = "NC_GLOBAL.", fixed = TRUE)
}
\arguments{
\item{NC}{a ncdf4 object}

\item{rm_pattern}{character a pattern of characters to remove from the
attribute names.  By default 'NC_GLOBAL.'.  Set to "" or NA to skip}

\item{fixed}{logical by default TRUE but see \code{grepl}}
}
\value{
named vector of global attributes
}
\description{
Retrieve a named list of global attributes
}

