% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{nctime_get}
\alias{nctime_get}
\title{Retrieve a vector of timestamps for a multilayer NC object or NULL otherwise}
\usage{
nctime_get(NC, name = "time", as_POSIXct = TRUE)
}
\arguments{
\item{NC}{a ncdf4 class object}

\item{name}{the name of the time-associated variable, by default 'time'}

\item{as_POSIXct}{logical, if TRUE then convert to POSIXct}
}
\value{
a numeric vector of timestamps (possibly POSIXct) or NULL if none
}
\description{
Retrieve a vector of timestamps for a multilayer NC object or NULL otherwise
}
