% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nc.R
\name{ncvardim_get}
\alias{ncvardim_get}
\title{Retrieve a list of dimension vectors, one for each variable}
\usage{
ncvardim_get(NC)
}
\arguments{
\item{NC}{a ncdf4 object}
}
\value{
a named list of variable dimension vectors
}
\description{
Retrieve a list of dimension vectors, one for each variable
}

