% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{ncvarname_get}
\alias{ncvarname_get}
\title{Retrieve a vector of variable names}
\usage{
ncvarname_get(NC)
}
\arguments{
\item{NC}{a ncdf4 object}
}
\value{
a named vector variable names
}
\description{
Retrieve a vector of variable names
}
