% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NHSCE.R
\docType{class}
\name{NHSCERefClass-class}
\alias{NHSCERefClass}
\alias{NHSCERefClass-class}
\title{A subclass of SPNCRefClass for NOAA/NCDC Climate Data Record of snow cover
extent}
\description{
Reference \url{https://climatedataguide.ucar.edu/climate-data/snow-cover-extent-northern-hemisphere-climate-data-record-rutgers}
}
\details{
Source \url{http://www.ncdc.noaa.gov/thredds/dodsC/cdr/snowcover/nhsce_v01r01_19661004_latest.nc}

See \url{http://climate.rutgers.edu/snowcover/docs.php?target=vis}

See \url{http://www1.ncdc.noaa.gov/pub/data/sds/cdr/docs/NH-SCE-CDR-CATBD-Rev2.pdf}
}

