% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OISST.R
\name{OISST_get_raster}
\alias{OISST_get_raster}
\title{Get a raster for OISSTRefClass}
\usage{
OISST_get_raster(NC, what = NC$VARS[1], layer = 1, bb = NC$BB,
  crs = "+proj=longlat +datum=WGS84", flip = TRUE, time_fmt = "D\%Y\%j")
}
\arguments{
\item{NC}{OISSTRefClass object}

\item{what}{character one or more variable names or variable indices}

\item{layer}{numeric vector either a 1-based indices or POSIXct timestamps}

\item{bb}{a 4 element bounding box vector [left, right, bottom, top], defaults
to [-180, 180, -90, 90]}

\item{crs}{character, the coordinate reference system to apply}

\item{flip}{logical if TRUE then flip the raster in the y direction}

\item{time_fmt}{if multiple time layers are returned, this controls the layer names}
}
\value{
a \code{raster::brick} or \code{raster::layer} object or NULL
}
\description{
Get a raster for OISSTRefClass
}

