% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OISST.R
\docType{class}
\name{OISSTRefClass-class}
\alias{OISSTRefClass}
\alias{OISSTRefClass-class}
\title{A subclass of SPNCRefClass for NOAA Optimum Interpolation (OI) Sea Surface Temperature \url{http://www.esrl.noaa.gov/psd/data/gridded/data.noaa.oisst.v2.html}}
\description{
A subclass of SPNCRefClass for NOAA Optimum Interpolation (OI) Sea Surface Temperature \url{http://www.esrl.noaa.gov/psd/data/gridded/data.noaa.oisst.v2.html}
}

