% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OISST.R
\name{OISSTRefClass_get_raster}
\alias{OISSTRefClass_get_raster}
\title{Get a raster}
\arguments{
\item{what}{character one or more variable names or variable indices}

\item{time}{numeric vector either a 1-based indices or POSIXct timestamps}

\item{bb}{a 4 element bounding box vector [left, right, bottom, top], defaults
to [-180, 180, -90, 90]}

\item{crs}{character, the coordiante reference system to apply}

\item{flip}{logical if TRUE then flip the raster in the y direction}
}
\value{
a \code{raster::brick} or \code{raster::layer} object or NULL
}
\description{
Requests can be made for multiple variables at multiple times.  In such cases
it is impotant to know that the order of rasters is each time for each variable
requested.  Thus requesting variables 'v1', 'v2', 'v3' at times t1 and t2
results in a stack with v1-t1, v1-t2, v2-t1, v2-t2, v3-t1, v3-t2 order.
}
