% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OISST.R
\name{OISSTRefClass_step}
\alias{OISSTRefClass_step}
\title{Get the step size (resolution) as a guess - the space between the first 
on and lat values.  Subclasses with unambiguous steps sizes, like L3SMI and  
MURSST, should override this method.}
\value{
two element numeric vector of step size in x and y or NULL
These may be signed for descending values of lon and or lat
}
\description{
Get the step size (resolution) as a guess - the space between the first 
on and lat values.  Subclasses with unambiguous steps sizes, like L3SMI and  
MURSST, should override this method.
}
