% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OISST.R
\name{OISSTRefClass_time_index}
\alias{OISSTRefClass_time_index}
\title{Compute time indices (which must be contiguous) from POSIXt, Date or index.
An error is thrown if the dates requested are not contiguous.}
\arguments{
\item{when}{numeric, POSIXct or Date times}

\item{no_zero}{logical, any times requested before the first time are mapped
to the first time}
}
\value{
one or more indices
}
\description{
Compute time indices (which must be contiguous) from POSIXt, Date or index.
An error is thrown if the dates requested are not contiguous.
}

