% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SPNC-package.R
\docType{package}
\name{spnc-package}
\alias{spnc-package}
\title{A simplified interface between spatial classes and NetCDF data sources.}
\description{
A simplified interface between spatial classes and NetCDF data sources.
}

