% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SPNC.R
\name{SPNC}
\alias{SPNC}
\title{A function to create an SPNCRefClass or subclass reference}
\usage{
SPNC(nc, bb = c(-180, 180, -90, 90), ...)
}
\arguments{
\item{nc}{'ncdf4' class object or path to one}

\item{bb}{a 4 element bounding box vector [left, right, bottom, top], defaults
to [-180, 180, -90, 90]}

\item{...}{futher arguments}
}
\value{
a SPNCRefClass object or subclass or NULL
}
\description{
A function to create an SPNCRefClass or subclass reference
}

