% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{spnc_flavor}
\alias{spnc_flavor}
\title{Determine the type and source of the netcdf data}
\usage{
spnc_flavor(x)
}
\arguments{
\item{x}{the netcdf4 or SPNCRefClass object}
}
\value{
return list of 
 \itemize{
   \item source character of 'MURSST', 'OISST', etc. "" means unknown
   \item type character of 'raster', 'points', etc. "" means unknown
   \item local logical (or NA) flag to indicate local or http source. NA mean unknown
 }
}
\description{
Determine the type and source of the netcdf data
}

