% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SPNC.R
\docType{class}
\name{SPNCRefClass-class}
\alias{SPNCRefClass}
\alias{SPNCRefClass-class}
\title{An object that connects to NetCDF (file or OpeNDAP) using ncdf4}
\description{
An object that connects to NetCDF (file or OpeNDAP) using ncdf4
}
\section{Fields}{

\describe{
\item{\code{flavor}}{a character vector of at least 'source' and 'type'}

\item{\code{the}}{ncdf4 class object}

\item{\code{the}}{4 element bounding box}

\item{\code{the}}{dimensions}

\item{\code{the}}{variable names}

\item{\code{possibly}}{a vector of lons}

\item{\code{possibly}}{a vector of lats}

\item{\code{possibly}}{a vector of z-values}

\item{\code{possibly}}{a vector of times}
}}

