% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SPNC.R
\name{SPNCRefClass_extent}
\alias{SPNCRefClass_extent}
\title{Compute extent given a bounding box}
\arguments{
\item{bb}{the bounding box needed, if missing the current one is used}
}
\value{
a raster::Extent object
}
\description{
Compute extent given a bounding box
}

