% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SPNC.R
\name{SPNCRefClass_get_dims}
\alias{SPNCRefClass_get_dims}
\title{Retrieve a named vector of dimensions}
\value{
a named numeric vector of dimensions or NULL
}
\description{
Retrieve a named vector of dimensions
}

