% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPNC.R
\name{SPNCRefClass_get_global_atts}
\alias{SPNCRefClass_get_global_atts}
\title{Get global attributes}
\arguments{
\item{...}{further arguments for \code{ncglobal_atts}}
}
\value{
a named list of global attributes or NULL
}
\description{
Get global attributes
}

