% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SPNC.R
\name{SPNCRefClass_get_path}
\alias{SPNCRefClass_get_path}
\title{Get path (line) data}
\arguments{
\item{what}{character one or more variable names or variable indices}

\item{crs}{character, the coordiante reference system to apply}
}
\value{
a \code{SpatialLinesDataFrame} or NULL
}
\description{
Get path (line) data
}

