% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPNC.R
\name{SPNCRefClass_get_points}
\alias{SPNCRefClass_get_points}
\title{Get point data}
\arguments{
\item{what}{character one or more variable names or variable indices}

\item{crs}{character, the coordiante reference system to apply}
}
\value{
a \code{SpatialPointsDataFrame} or NULL
}
\description{
Get point data
}
