% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SPNC.R
\name{SPNCRefClass_open}
\alias{SPNCRefClass_open}
\title{Open the ncdf4 object}
\arguments{
\item{path}{the path to the connection, if not present then try the
object's path}
}
\value{
logical
}
\description{
#'
}

