% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPNC.R
\name{SPNCRefClass_order_subset}
\alias{SPNCRefClass_order_subset}
\title{Order a subset coordinate system into the order matching that stored in
the ncdf file.  For ncdf4 the order should be X,Y,Z,T (which might be 
lon, lat, zlev, time, but not necessarily those names).  Honestly, this
probably should not even be a method - but use at your own risk/pain.}
\arguments{
\item{s}{a subset list comprised of at least start and count}
}
\value{
the same list with elements of start and count properly ordered
}
\description{
Order a subset coordinate system into the order matching that stored in
the ncdf file.  For ncdf4 the order should be X,Y,Z,T (which might be 
lon, lat, zlev, time, but not necessarily those names).  Honestly, this
probably should not even be a method - but use at your own risk/pain.
}

