% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SPNC.R
\name{SPNCRefClass_subset_bbox}
\alias{SPNCRefClass_subset_bbox}
\title{Craft subset indices into a ncdf array}
\arguments{
\item{bb}{numeric, four element bounding box [left, right, bottom, top]}
}
\value{
a list of \code{start} indices in x and y, \code{counts} in x and y and
   a possibly updated copy of \code{bb} vector of [left, right, bottom, top]
}
\description{
Craft subset indices into a ncdf array
}
\seealso{
\url{https://stat.ethz.ch/pipermail/r-help/2011-March/272641.html}
}

