% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPNC.R
\name{SPNCRefClass_time_index}
\alias{SPNCRefClass_time_index}
\title{Compute time indices (which must be contiguous) from POSIXt, Date or index.
An warning is thrown if the dates requested are not contiguous.  If the object
does not have TIME element then 1 is returned.}
\arguments{
\item{when}{numeric, POSIXct or Date times}

\item{no_zero}{logical, any times requested before the first time are mapped
to the first time}
}
\value{
one or more indices
}
\description{
Compute time indices (which must be contiguous) from POSIXt, Date or index.
An warning is thrown if the dates requested are not contiguous.  If the object
does not have TIME element then 1 is returned.
}

