% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{subset_nav}
\alias{subset_nav}
\title{Craft subset indices into a ncdf array}
\usage{
subset_nav(bb = NULL, lon = c(-180, 180), lat = c(-90, 90))
}
\arguments{
\item{bb}{numeric, four element bounding box [left, right, bottom, top]}

\item{lon}{numeric vector of lons to select from - these come from the LON attribute in nc}

\item{lat}{numeric vector of lats to select from - these come from the LAT attribute in nc}
}
\value{
a list of \code{start} indices in x and y, \code{counts} in x and y and
   a possibly updated copy of \code{bb} vector of [left, right, bottom, top]
}
\description{
Craft subset indices into a ncdf array
}
\seealso{
\url{http://r.789695.n4.nabble.com/Lat-Lon-NetCDF-subset-td3394326.html}
}

