% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{subset_points}
\alias{subset_points}
\title{Converts Lon/Lat to start and count suitable for extracting points}
\usage{
subset_points(x, y = NULL, lon = c(-180, 180), lat = c(-90, 90))
}
\arguments{
\item{x}{either a vector of longitude or a 2d array [lon,lat] or a data.frame
with lon,lat columns}

\item{y}{numeric vector, if x is a vector, otherwise NULL}

\item{lon}{numeric vector of lons to select from - these come from the LON attribute in nc}

\item{lat}{numeric vector of lats to select from - these come from the LAT attribute in nc}
}
\value{
a list of start/count elements, one per xy input.  For inputs that
   fall beyond the bounds of the data then NA is returned.
}
\description{
Converts Lon/Lat to start and count suitable for extracting points
}

