% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{to180}
\alias{to180}
\title{Convert [0,360] longitudes to [-180, 180]}
\usage{
to180(x)
}
\arguments{
\item{x}{numeric vector, no check is done for being withing 0,360 range}
}
\value{
numeric vector
}
\description{
Convert [0,360] longitudes to [-180, 180]
}
