% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{to180BB}
\alias{to180BB}
\title{Convert bounding box [0,360] longitudes to [-180, 180]}
\usage{
to180BB(x)
}
\arguments{
\item{x}{numeric bounding box vector, no check is done for being withing 0,360 range}
}
\value{
numeric bounding box vector
}
\description{
Bounding boxes are 4 element vectors of [left, right, bottom, top]
}
