% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{to360}
\alias{to360}
\title{Convert [-180,180] longitudes to [0,360]}
\usage{
to360(x)
}
\arguments{
\item{x}{numeric vector, no check is done for being withing 0,360 range}
}
\value{
numeric vector
}
\description{
Convert [-180,180] longitudes to [0,360]
}
