% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{to360BB}
\alias{to360BB}
\title{Convert [-180,180] bounding box longitudes to [0,360]}
\usage{
to360BB(x)
}
\arguments{
\item{x}{numeric bounding box vector, no check is done for being withing 0,360 range}
}
\value{
numeric bounding box vector
}
\description{
Bounding boxes are 4 element vectors of [left, right, bottom, top]
}

