% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{get_res}
\alias{get_res}
\title{Get a 2 element vector of [dx, dy] resolution}
\usage{
get_res(x)
}
\arguments{
\item{x}{netcd4 or raster object}
}
\value{
2 element numeric of resolution in [x,y]
}
\description{
Get a 2 element vector of [dx, dy] resolution
}
