% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3SMI.R
\name{L3SMI_get_points}
\alias{L3SMI_get_points}
\title{Get points for L3SMIRefClass}
\usage{
L3SMI_get_points(NC, x, y = NULL, what = NC$VARS[1])
}
\arguments{
\item{NC}{L3SMIRefClass object}

\item{x}{numeric, vector of longitude points or, if y is NULL, a matrix [x,y] 
or a data.frame or list with x,y elements}

\item{y}{numeric vector of lattitude points or NULL}
}
\value{
numeric vector of values
}
\description{
Get points for L3SMIRefClass
}

