% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{nav_from_bb}
\alias{nav_from_bb}
\title{Generate the navigation elements required to subset an ncdf4 object.
Returns a list of start indices and count and an adjusted bounding box}
\usage{
nav_from_bb(r, bb)
}
\arguments{
\item{r}{raster object}

\item{bb}{a 4 element vector describing the bounding box [left, right, bottom, top]}
}
\value{
a 3 element vector of start [ix,iy], count, [nx, ny], and
   an adjusted bb [left, right, bottom, top] and ext (extent) same as bb
}
\description{
Generate the navigation elements required to subset an ncdf4 object.
Returns a list of start indices and count and an adjusted bounding box
}
