% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OISST.R
\docType{class}
\name{OISSTRefClass-class}
\alias{OISSTRefClass}
\alias{OISSTRefClass-class}
\title{A subclass of SPNCRefClass for NOAA Optimum Interpolation (OI) Sea Surface Temperature \url{http://www.esrl.noaa.gov/psd/data/gridded/data.noaa.oisst.v2.html}}
\description{
Lon and Lat appear to be cell centers with 0.25 x 0.25 degree resolution
Lat is an ascending order [-90, 90] and while Lon in mapped to [0,360] 
bounding box requests must follow the [0,360] form.
}

