% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OISST.R
\name{OISSTRefClass_get_raster}
\alias{OISSTRefClass_get_raster}
\title{Get a raster}
\arguments{
\item{what}{character one or more variable names or variable indices}

\item{layer}{numeric vector either a 1-based indices or POSIXct timestamps}

\item{crs}{character, the coordiante reference system to apply}

\item{flip}{logical if TRUE then flip the raster in the y direction}

\item{time_fmt}{if multiple time layers are returned, this controls the layer names}
}
\value{
a \code{raster::brick} or \code{raster::layer} object or NULL
}
\description{
Get a raster
}

