% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPNC.R
\name{SPNC}
\alias{SPNC}
\title{A function to create an SPNCRefClass or subclass reference}
\usage{
SPNC(nc, bb = NULL, nc_verbose = FALSE, n_tries = 3, ...)
}
\arguments{
\item{nc}{'ncdf4' class object or path to one}

\item{bb}{a 4 element bounding box vector [left, right, bottom, top], defaults
to NULL}

\item{nc_verbose}{logical, passed to ncdf4::nc_open}

\item{n_tries}{numeric, when nc is a path we try to open upt to n_tries before failing
This helps accomodate occasional network/server issues.}

\item{...}{futher arguments}
}
\value{
a SPNCRefClass object or subclass or NULL
}
\description{
A function to create an SPNCRefClass or subclass reference
}
