% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{spnc_flavor}
\alias{spnc_flavor}
\title{Determine the type and source of the netcdf data}
\usage{
spnc_flavor(x)
}
\arguments{
\item{x}{the netcdf4 or SPNCRefClass object}

\item{return}{character vector of source and type.  If each is "" then no flavor determined.}
}
\description{
Determine the type and source of the netcdf data
}

