% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPNC.R
\name{SPNCRefClass_open}
\alias{SPNCRefClass_open}
\title{Open the ncdf4 object}
\arguments{
\item{path}{the path to the connection, if not present then try the 
object's path}

\item{...}{furtehr arguments for ncfd4::nc_open}
}
\value{
logical
}
\description{
Open the ncdf4 object
}
